+++ test/index.js
@@ -4,6 +4,7 @@ const { TestDirector } = require('test-director')
 
 const tests = new TestDirector()
 
+require('./lib/GraphQLUpload.test')(tests)
 require('./lib/graphqlUploadExpress.test')(tests)
 require('./lib/graphqlUploadKoa.test')(tests)
 require('./lib/ignoreStream.test')(tests)